# docs/openapi-mobile-v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Q-App Mobile API",
    "version": "v1",
    "description": "Machine-readable mobile-facing HTTP contract for Q-App v1."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:4000/v1",
      "description": "Local development"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Meta": {
        "type": "object",
        "required": ["requestId"],
        "properties": {
          "requestId": { "type": "string" },
          "count": { "type": "integer" }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "required": ["error", "code", "meta"],
        "properties": {
          "error": { "type": "string" },
          "code": { "type": "string" },
          "details": {},
          "meta": { "$ref": "#/components/schemas/Meta" }
        }
      },
      "BusinessGroupRef": {
        "type": "object",
        "required": ["id", "name", "slug"],
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "slug": { "type": "string" }
        }
      },
      "BusinessLocationRef": {
        "type": "object",
        "required": ["id", "name", "slug"],
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "slug": { "type": "string" }
        }
      },
      "Membership": {
        "type": "object",
        "required": ["role", "businessGroup"],
        "properties": {
          "role": { "type": "string" },
          "businessGroup": { "$ref": "#/components/schemas/BusinessGroupRef" }
        }
      },
      "StaffProfile": {
        "type": "object",
        "required": ["id", "displayName", "businessLocation"],
        "properties": {
          "id": { "type": "string" },
          "displayName": { "type": "string" },
          "businessLocation": { "$ref": "#/components/schemas/BusinessLocationRef" }
        }
      },
      "AuthUser": {
        "type": "object",
        "required": ["id", "appRole", "firstName", "email", "phone", "memberships", "staffProfiles"],
        "properties": {
          "id": { "type": "string" },
          "appRole": { "type": "string" },
          "firstName": { "type": "string" },
          "lastName": { "type": ["string", "null"] },
          "email": { "type": ["string", "null"] },
          "phone": { "type": ["string", "null"] },
          "memberships": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Membership" }
          },
          "staffProfiles": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/StaffProfile" }
          }
        }
      },
      "AuthTokens": {
        "type": "object",
        "required": [
          "accessToken",
          "refreshToken",
          "sessionId",
          "accessTokenExpiresInSeconds",
          "refreshTokenExpiresAt"
        ],
        "properties": {
          "accessToken": { "type": "string" },
          "refreshToken": { "type": "string" },
          "sessionId": { "type": "string" },
          "accessTokenExpiresInSeconds": { "type": "integer" },
          "refreshTokenExpiresAt": { "type": "string", "format": "date-time" }
        }
      },
      "AuthLoginPayload": {
        "type": "object",
        "required": ["user", "tokens"],
        "properties": {
          "user": { "$ref": "#/components/schemas/AuthUser" },
          "tokens": { "$ref": "#/components/schemas/AuthTokens" }
        }
      },
      "Session": {
        "type": "object",
        "required": ["id", "createdAt", "expiresAt", "revokedAt", "lastUsedAt", "isCurrent"],
        "properties": {
          "id": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" },
          "expiresAt": { "type": "string", "format": "date-time" },
          "revokedAt": { "type": ["string", "null"], "format": "date-time" },
          "lastUsedAt": { "type": ["string", "null"], "format": "date-time" },
          "userAgent": { "type": ["string", "null"] },
          "ipAddress": { "type": ["string", "null"] },
          "deviceName": { "type": ["string", "null"] },
          "platform": { "type": ["string", "null"] },
          "isCurrent": { "type": "boolean" }
        }
      },
      "AuthActivityActor": {
        "type": ["object", "null"],
        "properties": {
          "id": { "type": "string" },
          "email": { "type": ["string", "null"] },
          "firstName": { "type": "string" },
          "lastName": { "type": ["string", "null"] }
        }
      },
      "AuthActivityUser": {
        "type": ["object", "null"],
        "properties": {
          "id": { "type": "string" },
          "email": { "type": ["string", "null"] },
          "phone": { "type": ["string", "null"] },
          "firstName": { "type": "string" },
          "lastName": { "type": ["string", "null"] }
        }
      },
      "AuthActivityEvent": {
        "type": "object",
        "required": ["id", "createdAt", "type", "identifier", "metadata", "sessionId", "user", "actorUser"],
        "properties": {
          "id": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" },
          "type": { "type": "string" },
          "identifier": { "type": ["string", "null"] },
          "ipAddress": { "type": ["string", "null"] },
          "userAgent": { "type": ["string", "null"] },
          "deviceName": { "type": ["string", "null"] },
          "platform": { "type": ["string", "null"] },
          "metadata": {},
          "sessionId": { "type": ["string", "null"] },
          "user": { "$ref": "#/components/schemas/AuthActivityUser" },
          "actorUser": { "$ref": "#/components/schemas/AuthActivityActor" }
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Bad request",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": {
              "error": "Invalid payload.",
              "code": "BAD_REQUEST",
              "meta": { "requestId": "request-id" }
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Unauthorized",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": {
              "error": "Authentication required.",
              "code": "UNAUTHORIZED",
              "meta": { "requestId": "request-id" }
            }
          }
        }
      },
...
      "NotFound": {
        "description": "Not found",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": {
              "error": "Resource not found.",
              "code": "NOT_FOUND",
              "meta": { "requestId": "request-id" }
            }
          }
        }
      },
      "Conflict": {
        "description": "Conflict",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": {
              "error": "Conflict.",
              "code": "CONFLICT",
              "meta": { "requestId": "request-id" }
            }
          }
        }
      },
      "Locked": {
        "description": "Account locked",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": {
              "error": "Account is temporarily locked due to repeated failed login attempts.",
              "code": "ACCOUNT_LOCKED",
              "details": {
                "lockedUntil": "2026-06-12T10:30:00.000Z"
              },
              "meta": { "requestId": "request-id" }
            }
          }
        }
      },
      "RateLimited": {
        "description": "Rate limited",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": {
              "error": "Too many login attempts. Please try again later.",
              "code": "LOGIN_RATE_LIMITED",
              "meta": { "requestId": "request-id" }
            }
          }
        }
      }
    }
  },
  "paths": {
    "/auth/login": {
      "post": {
        "tags": ["Auth"],
        "operationId": "authLogin",
        "summary": "Login with identifier and password",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["identifier", "password"],
                "properties": {
                  "identifier": { "type": "string" },
                  "password": { "type": "string" },
                  "deviceName": { "type": "string" },
                  "platform": { "type": "string" }
                }
              },
              "example": {
                "identifier": "owner@fadeyard.demo",
                "password": "QappOwner123!",
                "deviceName": "iPhone 16 Pro",
                "platform": "ios"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Authenticated session created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["data", "item", "meta"],
                  "properties": {
                    "data": { "$ref": "#/components/schemas/AuthLoginPayload" },
                    "item": {},
                    "meta": { "$ref": "#/components/schemas/Meta" }
                  }
                },
                "example": {
                  "data": {
                    "user": {
                      "id": "user-id",
                      "appRole": "USER",
                      "firstName": "Owen",
                      "lastName": "Owner",
                      "email": "owner@fadeyard.demo",
                      "phone": "+447400000202",
                      "memberships": [],
                      "staffProfiles": []
                    },
                    "tokens": {
                      "accessToken": "<jwt>",
                      "refreshToken": "<refresh>",
                      "sessionId": "session-id",
                      "accessTokenExpiresInSeconds": 900,
                      "refreshTokenExpiresAt": "2026-07-12T09:15:30.000Z"
                    }
                  },
                  "item": {},
                  "meta": {
                    "requestId": "request-id"
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "423": { "$ref": "#/components/responses/Locked" },
          "429": { "$ref": "#/components/responses/RateLimited" }
        }
      }
    },
    "/auth/refresh": {
      "post": {
        "tags": ["Auth"],
        "operationId": "authRefresh",
        "summary": "Refresh an authenticated session",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["refreshToken"],
                "properties": {
                  "refreshToken": { "type": "string" }
                }
              },
              "example": {
                "refreshToken": "<refresh-token>"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Session refreshed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": { "$ref": "#/components/schemas/AuthLoginPayload" },
                    "item": {},
                    "meta": { "$ref": "#/components/schemas/Meta" }
                  }
                },
                "example": {
                  "data": {
                    "user": {
                      "id": "user-id",
                      "appRole": "USER",
                      "firstName": "Owen",
                      "lastName": "Owner",
                      "email": "owner@fadeyard.demo",
                      "phone": "+447400000202",
                      "memberships": [],
                      "staffProfiles": []
                    },
                    "tokens": {
                      "accessToken": "<jwt>",
                      "refreshToken": "<refresh>",
                      "sessionId": "session-id",
                      "accessTokenExpiresInSeconds": 900,
                      "refreshTokenExpiresAt": "2026-07-12T09:30:30.000Z"
                    }
                  },
                  "item": {},
                  "meta": { "requestId": "request-id" }
                }
              }
            }
          },
          "409": { "$ref": "#/components/responses/Conflict" }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "tags": ["Auth"],
        "operationId": "authLogout",
        "summary": "Logout by refresh token",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["refreshToken"],
                "properties": {
                  "refreshToken": { "type": "string" }
                }
              },
              "example": {
                "refreshToken": "<refresh-token>"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Session revoked",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "required": ["revoked"],
                      "properties": {
                        "revoked": { "type": "boolean" }
                      }
                    },
                    "item": {},
                    "meta": { "$ref": "#/components/schemas/Meta" }
                  }
                },
                "example": {
                  "data": { "revoked": true },
                  "item": {},
                  "meta": { "requestId": "request-id" }
                }
              }
            }
          }
        }
      }
    },
    "/auth/me": {
      "get": {
        "tags": ["Auth"],
        "operationId": "authMe",
        "summary": "Get current authenticated user",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Current user profile",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": { "$ref": "#/components/schemas/AuthUser" },
                    "item": {},
                    "meta": { "$ref": "#/components/schemas/Meta" }
                  }
                },
                "example": {
                  "data": {
                    "id": "user-id",
                    "appRole": "USER",
                    "firstName": "Owen",
                    "lastName": "Owner",
                    "email": "owner@fadeyard.demo",
                    "phone": "+447400000202",
                    "memberships": [],
                    "staffProfiles": []
                  },
                  "item": {},
                  "meta": { "requestId": "request-id" }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/auth/sessions": {
      "get": {
        "tags": ["Auth"],
        "operationId": "authSessions",
        "summary": "List active and historical sessions for the authenticated user",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Session list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Session" }
                    },
                    "items": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Session" }
                    },
                    "meta": { "$ref": "#/components/schemas/Meta" }
                  }
                },
                "example": {
                  "data": [
                    {
                      "id": "session-id",
                      "createdAt": "2026-06-12T09:15:30.000Z",
                      "expiresAt": "2026-07-12T09:15:30.000Z",
                      "revokedAt": null,
                      "lastUsedAt": "2026-06-12T09:15:30.000Z",
                      "userAgent": "QApp iOS/1.0",
                      "ipAddress": "::ffff:127.0.0.1",
                      "deviceName": "iPhone 16 Pro",
                      "platform": "ios",
                      "isCurrent": true
                    }
                  ],
                  "items": [],
                  "meta": {
                    "count": 1,
                    "requestId": "request-id"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/auth/activity": {
      "get": {
        "tags": ["Auth"],
        "operationId": "authActivity",
        "summary": "List recent authenticated user security activity",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": { "type": "integer", "minimum": 1, "maximum": 100 }
          }
        ],
        "responses": {
          "200": {
            "description": "Activity list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/AuthActivityEvent" }
                    },
                    "items": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/AuthActivityEvent" }
                    },
                    "meta": { "$ref": "#/components/schemas/Meta" }
                  }
                },
                "example": {
                  "data": [
                    {
                      "id": "event-id",
                      "createdAt": "2026-06-12T09:20:11.000Z",
                      "type": "LOGIN_SUCCEEDED",
                      "identifier": "owner@fadeyard.demo",
                      "ipAddress": "::ffff:127.0.0.1",
                      "userAgent": "QApp iOS/1.0",
                      "deviceName": "iPhone 16 Pro",
                      "platform": "ios",
                      "metadata": null,
                      "sessionId": "session-id",
                      "user": {
                        "id": "user-id",
                        "email": "owner@fadeyard.demo",
                        "phone": "+447400000202",
                        "firstName": "Owen",
                        "lastName": "Owner"
                      },
                      "actorUser": null
                    }
                  ],
                  "items": [],
                  "meta": { "count": 1, "requestId": "request-id" }
                }
              }
            }
          }
        }
      }
    },
    "/auth/logout-current": {
      "post": {
        "tags": ["Auth"],
        "operationId": "authLogoutCurrent",
        "summary": "Logout the current session",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Current session revoked",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "required": ["revoked"],
                      "properties": {
                        "revoked": { "type": "boolean" }
                      }
                    },
                    "item": {},
                    "meta": { "$ref": "#/components/schemas/Meta" }
                  }
                },
                "example": {
                  "data": { "revoked": true },
                  "item": {},
                  "meta": { "requestId": "request-id" }
                }
              }
            }
          }
        }
      }
    },
    "/auth/logout-other-sessions": {
      "post": {
        "tags": ["Auth"],
        "operationId": "authLogoutOtherSessions",
        "summary": "Logout every other session for the authenticated user",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Other sessions revoked",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "required": ["revoked", "revokedSessionsCount"],
                      "properties": {
                        "revoked": { "type": "boolean" },
                        "revokedSessionsCount": { "type": "integer" }
                      }
                    },
                    "item": {},
                    "meta": { "$ref": "#/components/schemas/Meta" }
                  }
                },
                "example": {
                  "data": {
                    "revoked": true,
                    "revokedSessionsCount": 1
                  },
                  "item": {},
                  "meta": { "requestId": "request-id" }
                }
              }
            }
          }
        }
      }
    },
    "/auth/sessions/{sessionId}/revoke": {
      "post": {
        "tags": ["Auth"],
        "operationId": "authRevokeSession",
        "summary": "Revoke one owned session",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Session revoked",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "required": ["revoked", "sessionId"],
                      "properties": {
                        "revoked": { "type": "boolean" },
                        "sessionId": { "type": "string" }
                      }
                    },
                    "item": {},
                    "meta": { "$ref": "#/components/schemas/Meta" }
                  }
                },
                "example": {
                  "data": {
                    "revoked": true,
                    "sessionId": "session-id"
                  },
                  "item": {},
                  "meta": { "requestId": "request-id" }
                }
              }
            }
          }
        }
      }
    },
    "/auth/password-reset/request": {
      "post": {
        "tags": ["Auth"],
        "operationId": "authPasswordResetRequest",
        "summary": "Request password reset",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["identifier"],
                "properties": {
                  "identifier": { "type": "string" }
                }
              },
              "example": {
                "identifier": "owner@fadeyard.demo"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Neutral password reset request response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "required": ["message"],
                      "properties": {
                        "message": { "type": "string" },
                        "resetTokenPreview": { "type": "string" },
                        "resetUrlPreview": { "type": "string" }
                      }
                    },
                    "item": {},
                    "meta": { "$ref": "#/components/schemas/Meta" }
                  }
                },
                "example": {
                  "data": {
                    "message": "If the account exists, a password reset link has been prepared."
                  },
                  "item": {},
                  "meta": { "requestId": "request-id" }
                }
              }
            }
          },
          "429": { "$ref": "#/components/responses/RateLimited" }
        }
      }
    },
    "/auth/password-reset/confirm": {
      "post": {
        "tags": ["Auth"],
        "operationId": "authPasswordResetConfirm",
        "summary": "Confirm password reset with token",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["token", "password"],
                "properties": {
                  "token": { "type": "string" },
                  "password": { "type": "string" }
                }
              },
              "example": {
                "token": "<reset-token>",
                "password": "QappOwnerReset123!"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password reset completed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "required": ["reset", "revokedSessions"],
                      "properties": {
                        "reset": { "type": "boolean" },
                        "revokedSessions": { "type": "boolean" }
                      }
                    },
                    "item": {},
                    "meta": { "$ref": "#/components/schemas/Meta" }
                  }
                },
                "example": {
                  "data": {
                    "reset": true,
                    "revokedSessions": true
                  },
                  "item": {},
                  "meta": { "requestId": "request-id" }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" }
        }
      }
    },
...
  }
}